% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTestData.R
\name{GenerateTestData}
\alias{GenerateTestData}
\title{Function to generate a test data}
\usage{
GenerateTestData(error_id, n, pnum, beta_star, gamma_star, causalrare,
  causalcommon, SKAT.haplotypes)
}
\arguments{
\item{error_id}{A number indicates error distribution: 1 for N(0,1); 2 for Chi square with df=2; 3 for Cauchy (0,1); 4 for t(df = 2); 5 for t(df = 3).}

\item{n}{Sample size.}

\item{pnum}{Number of variants.}

\item{beta_star}{coefficients for \code{X}.}

\item{gamma_star}{Control heteroscedastic case: \code{error = (1+gamma_star*X)*e}.}

\item{causalrare}{Causal ratio for rare variants.}

\item{causalcommon}{Causal ratio for common variants.}

\item{SKAT.haplotypes}{Data from SKAT pacakge.}
}
\value{
A list contains response \code{y} under null distribution, \code{y} under alternative distrubtion, \code{x} genotype, \code{c} covariates and \code{error}.
}
\description{
Function to generate a test data
}
