% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_iQRAT.R
\name{meta_iQRAT}
\alias{meta_iQRAT}
\title{Function for meta-analysis using Fisher's method}
\usage{
meta_iQRAT(pval.list)
}
\arguments{
\item{pval.list}{A list of independent p values. Each element is a vector with same length.}
}
\value{
The p value for meta-analysis.
}
\description{
This funciton uses Fisher's method to combine a list of independent p values and return the p value for mets-analysis.
}
