% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewRanks.R
\name{NewRanks}
\alias{NewRanks}
\title{This is the file according to quantreg, but adjusted for LiQRAT.
It takes the sparse matrix dsol[,1:(J-1)] - dsol[,2:J]}
\usage{
NewRanks(v, score = "wilcoxon", Delta = NULL)
}
\description{
This is the file according to quantreg, but adjusted for LiQRAT.
It takes the sparse matrix dsol[,1:(J-1)] - dsol[,2:J]
}
