% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Null_model.R
\name{Null_model}
\alias{Null_model}
\title{Fit the null model
This function provides null model fit. The default option is to use quantreg since it is faster. When
dataset is too large, quantreg will have memory limitation issue, and we will use the self-implemented
algorithm.}
\usage{
Null_model(Y, C)
}
\arguments{
\item{Y}{A vector of response, size \eqn{n}.}

\item{C}{A vector or matrix of covariates.}
}
\description{
Fit the null model
This function provides null model fit. The default option is to use quantreg since it is faster. When
dataset is too large, quantreg will have memory limitation issue, and we will use the self-implemented
algorithm.
}
