% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSignal.R
\name{plotiQRAT}
\alias{plotiQRAT}
\title{Visualize quantile effect for the entire process}
\usage{
plotiQRAT(Y, X, C)
}
\arguments{
\item{Y}{A vector of response, size \eqn{n}.}

\item{X}{A n by p matrix of genotype.}

\item{C}{A vector or matrix of covariates.}
}
\value{
A list contains quantile levels and corresponding p values.
}
\description{
This function implements the visualization tool as an extension of the main approach.
}
\examples{
library(quantreg)
library(SKAT)

error_id = 1

n= 500
pnum = 70
beta_star = 0.3
gamma_star = 0.1

causalrare = 0.3
causalcommon = 0.2

# select region


data("SKAT.haplotypes")
Data = GenerateTestData(error_id, n, pnum, beta_star, gamma_star,
                        causalrare, causalcommon, SKAT.haplotypes)
out = plotiQRAT(Data$y_a, Data$x, Data$c)
plot(out$taus, -log10(out$pval), type = "l", xlab = "quantiles", ylab = "-log10(p)")
}
