% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iQRAT.R
\name{iQRAT}
\alias{iQRAT}
\title{Integrated Quantile Rank Test for group-wise joint effect of rare and common variants in sequencing study}
\usage{
iQRAT(Y, X, C, cutoff = NULL, weight = TRUE, method.type = "S")
}
\arguments{
\item{Y}{A vector of response, size \eqn{n}.}

\item{X}{A n by p matrix of genotype.}

\item{C}{A vector or matrix of covariates.}

\item{cutoff}{A number for separating common and rare variants. If not assigned, the default cutoff is \eqn{1/\sqrt(2*n)}.}

\item{weight}{A logicalvariable indicating if different weights will be assigned to common and rare variants. The default is TRUE.}

\item{method.type}{A character, either "S" for \eqn{Q_S} or "B" for \eqn{Q_B}.}
}
\value{
A list contains p value, case and method type. Case indicates whether \code{X} is common, rare or mix.
}
\description{
This function implements the efficient quantile rank group-wise test, especially designed for testing the joint effect of rare
and common variants in sequencing study. Comparing to existing methods, it tests integrated associations across the entire response distribution, rather than
only focusing on mean or variability. This method does not have assumptions for error distributions. Thus, no quantile and rank normalization
is required for data pre-processing. Note that when applying to univariate, no weights will be assigned.
}
\section{References}{

Wang, T., Ionita-Laza, I. and Wei, Y. Efficient Quantile Rank Test for heterogeneous joint effect of rare and common variants in sequencing studies.
}

\examples{
library(quantreg)
library(SKAT)

error_id = 1

n= 500
pnum = 70
beta_star = 0.3
gamma_star = 0.1

causalrare = 0.3
causalcommon = 0.2

# select region


data("SKAT.haplotypes")
Data = GenerateTestData(error_id, n, pnum, beta_star, gamma_star,
                        causalrare, causalcommon, SKAT.haplotypes)

iQRAT(Data$y_a, Data$x, Data$c)




}
